fix: more updates for settings revamp
