+++ apps/extension/src/app/features/settings/settings-current.tsx
@@ -25,6 +25,7 @@ import { useModifierKey } from '@app/common/hooks/use-modifier-key';
 import { useWalletType } from '@app/common/use-wallet-type';
 import { Divider } from '@app/components/layout/divider';
 import { SignOut } from '@app/features/settings/sign-out/sign-out-confirm';
+import { useHasDefaultInMemoryWalletSecretKey } from '@app/store/in-memory-key/in-memory-key.selectors';
 import { useTogglePrivateMode } from '@app/store/settings/settings.actions';
 import { useIsPrivateMode } from '@app/store/settings/settings.selectors';
 
@@ -52,6 +53,76 @@ export function Settings({ canLockWallet = true }: SettingsProps) {
   const { isPressed: showAdvancedMenuOptions } = useModifierKey('alt', 120);
   const showLockWalletItem = canLockWallet && hasKeys && walletType === 'software';
   const showSignOutItem = hasKeys;
+  const hasDefaultInMemorySecretKey = useHasDefaultInMemoryWalletSecretKey();
+  const isWalletUnlocked = hasDefaultInMemorySecretKey || walletType === 'ledger';
+
+  const settingsItem = (
+    <DropdownMenu.Item
+      data-testid={SettingsSelectors.SettingsMenuItem}
+      onSelect={() => {
+        analytics.track('click_settings_menu_item');
+        void navigate(RouteUrls.Settings);
+      }}
+    >
+      <Flag img={<SettingsGearIcon />} width="100%">
+        <Flex justifyContent="space-between" textStyle="label.02">
+          Settings
+        </Flex>
+      </Flag>
+    </DropdownMenu.Item>
+  );
+
+  const togglePrivacyItem = (
+    <DropdownMenu.Item
+      data-testid={SettingsSelectors.TogglePrivacy}
+      onSelect={e => {
+        analytics.track('click_toggle_privacy');
+        togglePrivateMode();
+        e.preventDefault();
+      }}
+    >
+      <Flag img={isPrivateMode ? <Eye1ClosedIcon /> : <Eye1Icon />} width="100%">
+        <Flex justifyContent="space-between" textStyle="label.02">
+          Toggle privacy
+          <Switch.Root checked={isPrivateMode}>
+            <Switch.Thumb />
+          </Switch.Root>
+        </Flex>
+      </Flag>
+    </DropdownMenu.Item>
+  );
+
+  const lockWalletItem = (
+    <DropdownMenu.Item
+      onSelect={() => {
+        analytics.track('lock_session');
+        void lockWallet({
+          afterLock: () => navigate(RouteUrls.Unlock, { state: { from: location.pathname } }),
+        });
+      }}
+      data-testid={SettingsSelectors.LockListItem}
+    >
+      <Flag img={<LockIcon />} textStyle="label.02" width="100%">
+        Lock
+      </Flag>
+    </DropdownMenu.Item>
+  );
+
+  const signOutItem = (
+    <DropdownMenu.Item
+      onSelect={() => setShowSignOut(!showSignOut)}
+      data-testid={SettingsSelectors.SignOutListItem}
+    >
+      <Flag
+        color="red.action-primary-default"
+        img={<ExitIcon color="red.action-primary-default" />}
+        textStyle="label.02"
+        width="100%"
+      >
+        Sign out
+      </Flag>
+    </DropdownMenu.Item>
+  );
 
   return (
     <>
@@ -72,75 +143,17 @@ export function Settings({ canLockWallet = true }: SettingsProps) {
             })}
           >
             <DropdownMenu.Group>
-              <DropdownMenu.Item
-                data-testid={SettingsSelectors.SettingsMenuItem}
-                onSelect={() => {
-                  analytics.track('click_settings_menu_item');
-                  void navigate(RouteUrls.Settings);
-                }}
-              >
-                <Flag img={<SettingsGearIcon />} width="100%">
-                  <Flex justifyContent="space-between" textStyle="label.02">
-                    Settings
-                  </Flex>
-                </Flag>
-              </DropdownMenu.Item>
-              <DropdownMenu.Item
-                data-testid={SettingsSelectors.TogglePrivacy}
-                onSelect={e => {
-                  analytics.track('click_toggle_privacy');
-                  togglePrivateMode();
-                  e.preventDefault();
-                }}
-              >
-                <Flag img={isPrivateMode ? <Eye1ClosedIcon /> : <Eye1Icon />} width="100%">
-                  <Flex justifyContent="space-between" textStyle="label.02">
-                    Toggle privacy
-                    <Switch.Root checked={isPrivateMode}>
-                      <Switch.Thumb />
-                    </Switch.Root>
-                  </Flex>
-                </Flag>
-              </DropdownMenu.Item>
-              {(showAdvancedMenuOptions || showLockWalletItem || showSignOutItem) && (
-                <>
-                  <Divider />
-                  {showAdvancedMenuOptions && <AdvancedMenuItems />}
+              {isWalletUnlocked && settingsItem}
+              {isWalletUnlocked && togglePrivacyItem}
 
-                  {showLockWalletItem && (
-                    <DropdownMenu.Item
-                      onSelect={() => {
-                        analytics.track('lock_session');
-                        void lockWallet({
-                          afterLock: () =>
-                            navigate(RouteUrls.Unlock, { state: { from: location.pathname } }),
-                        });
-                      }}
-                      data-testid={SettingsSelectors.LockListItem}
-                    >
-                      <Flag img={<LockIcon />} textStyle="label.02" width="100%">
-                        Lock
-                      </Flag>
-                    </DropdownMenu.Item>
-                  )}
+              {(showAdvancedMenuOptions || showLockWalletItem || showSignOutItem) &&
+                isWalletUnlocked && <Divider />}
 
-                  {showSignOutItem && (
-                    <DropdownMenu.Item
-                      onSelect={() => setShowSignOut(!showSignOut)}
-                      data-testid={SettingsSelectors.SignOutListItem}
-                    >
-                      <Flag
-                        color="red.action-primary-default"
-                        img={<ExitIcon color="red.action-primary-default" />}
-                        textStyle="label.02"
-                        width="100%"
-                      >
-                        Sign out
-                      </Flag>
-                    </DropdownMenu.Item>
-                  )}
-                </>
-              )}
+              {showAdvancedMenuOptions && <AdvancedMenuItems />}
+
+              {showLockWalletItem && lockWalletItem}
+
+              {showSignOutItem && signOutItem}
             </DropdownMenu.Group>
           </DropdownMenu.Content>
         </DropdownMenu.Portal>
