+++ apps/extension/src/app/pages/onboarding/sign-in/forgot-password.tsx
@@ -3,7 +3,7 @@ import { EnterMnemonic } from './enter-mnemonic';
 export function ForgotPassword() {
   return (
     <EnterMnemonic
-      title={<>Forgot password?</>}
+      title="Forgot password?"
       description="No worries, sign in with your Secret Keys and restore your wallet."
     />
   );
