+++ apps/extension/src/app/pages/onboarding/sign-in/sign-in.tsx
@@ -3,11 +3,7 @@ import { EnterMnemonic } from './enter-mnemonic';
 export function SignIn() {
   return (
     <EnterMnemonic
-      title={
-        <>
-          Sign in <br /> with your <br /> Secret Key
-        </>
-      }
+      title="Sign in with your Secret Key"
       description="Speed things up by pasting your entire Secret Key in one go."
     />
   );
