+++ apps/extension/src/app/pages/view-secret-key/view-secret-key-current.tsx
@@ -9,10 +9,13 @@ import { Content } from '@app/components/layout';
 import { Header } from '@app/components/layout/headers/header';
 import { HeaderBackButton } from '@app/components/layout/headers/header-back-button';
 import { HeaderGrid } from '@app/components/layout/headers/header-grid';
+import {
+  DescriptionColumn,
+  TwoColumnLayout,
+} from '@app/components/layout/layouts/two-column-current.layout';
 import { RequestPassword } from '@app/components/request-password';
 import { useDefaultWalletSecretKey } from '@app/store/in-memory-key/in-memory-key.selectors';
-
-import { SecretKey } from './components/secret-key';
+import { SecretKey } from '@app/ui/components/secret-key-current/secret-key';
 
 export function ViewSecretKey() {
   const defaultWalletSecretKey = useDefaultWalletSecretKey();
@@ -33,7 +36,17 @@ export function ViewSecretKey() {
       <>
         {header}
         <Content>
-          <SecretKey secretKey={defaultWalletSecretKey ?? ''} />
+          <TwoColumnLayout
+            leftColumn={
+              <DescriptionColumn
+                title="Secret Key"
+                description="These 24 words are your Secret Key. They create your account, and you sign in on
+            different devices with them. Make sure to save these somewhere safe. If you lose these
+            words, you lose your account."
+              />
+            }
+            rightColumn={<SecretKey secretKey={defaultWalletSecretKey ?? ''} />}
+          />
         </Content>
       </>
     );
