+++ apps/extension/src/app/ui/components/secret-key-current/secret-key-word.tsx
